Remove GenerateValueOnAdd from fluent API
The problem with this is that it is really an internal detail which should never be switched off even when explicit key values are being used. Occasionally it cab be useful to switch it on for non-key, non-discriminator properties, but this would not be a common case and dropping down to core metadata is fine for this.

Also, removed it from the model snapshot since it really has no impact on store generation--StoreGeneratedPattern is for that. Removing it from the snapshot makes diffs easier to handle.

Also, did a name change that got missed when the reat of the metadata API review changes were applied.